+++ src/app.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { HealthCheckModule } from './modules/healthCheck/healthCheck.module';
+import { HealthCheckModule } from './modules/health-check/health-check.module';
 import { typeORMConfig } from './config/typeorm.config';
 import { TasksModule } from './modules/tasks/tasks.module';
 import { ProjectsModule } from './modules/projects/projects.module';
 import { CacheModule } from '@nestjs/cache-manager';
 import { redisStore } from 'cache-manager-redis-store';
 import type { RedisClientOptions } from 'redis';
-import { PersonalRecallsModule } from './modules/personalRecalls/personalRecalls.module';
+import { PersonalRecallsModule } from './modules/personal-recalls/personal-recalls.module';
 import { AuthModule } from './modules/auth/auth.module';
 import { UsersModule } from './modules/users/users.module';
 import { APP_GUARD } from '@nestjs/core';
