+++ src/common/enums/application-type.enum.ts
