+++ src/modules/mappings/user-projects/dtos/user-with-task.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { UserProject } from 'src/modules/mappings/userProjects/userProjects.entity';
+import { UserProject } from 'src/modules/mappings/user-projects/userProjects.entity';
 
 export class UserWithTasksDto {
     @ApiProperty()
