+++ src/modules/master-portfolios/master-portfolios.entity.ts
@@ -2,7 +2,7 @@ import { BaseEntity } from 'src/common/entities/base.entity';
 import { Column, Entity, ManyToOne } from 'typeorm';
 import { User } from '../users/entities/users.entity';
 import { Project } from '../projects/entities/projects.entity';
-import { portfolioType } from 'src/common/enums/portfolioType.enum';
+import { portfolioType } from 'src/common/enums/portfolio-type.enum';
 
 @Entity()
 export class MasterPortfolio extends BaseEntity {
