+++ src/modules/personal-recalls/personal-recalls.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { PersonalRecall } from './entities/personalRecalls.entity';
+import { PersonalRecall } from './entities/personal-recalls.entity';
 import { Repository } from 'typeorm';
-import { PersonalRecallResponseDto } from './dtos/personalRecall.dto';
+import { PersonalRecallResponseDto } from './dtos/personal-recall.dto';
 
 @Injectable()
 export class PersonalRecallsService {
