+++ src/modules/tasks/tasks.entity.ts
@@ -3,7 +3,7 @@ import { Column, Entity, ManyToOne, OneToMany } from 'typeorm';
 import { Comment } from '../comments/comments.entity';
 import { BaseEntity } from 'src/common/entities/base.entity';
 import { Manager } from '../mappings/managers/managers.entity';
-import { TaskFile } from '../mappings/taskFiles/taskFiles.entity';
+import { TaskFile } from '../mappings/task-files/task-files.entity';
 import { Step } from '../steps/steps.entity';
 
 @Entity()
